Implements VideoAttachment+WordPress
